+++ monitor-sdk/src/common/injectFetch.ts
