+++ monitor-sdk/src/common/injectXHR.ts
